[lono-inline] test mode runs all tests. Py merge*
Combine the walk, merge, and file compare tests into a single "tests"
mode. The compare tests are given as a set of before after files to
load/compare. The mergewith and deepmerge functions are added to the
python side along with the respective test code.

Keep a compare rule that allows manually specifying a before and after
file to test against.