[lono-inline] no global defs. Reduce python size.
Instead of doing evaluating the JSEval/PyEval code in a global
context, evaluate the code with the ns map as the context. And rename
the evaluated code section to JSInit and PyInit. A new exports
variable is defined for the JSInit section and any exposed
values/functions should be added to exports (which is just the ns
variable but passed in as exports by JS convention). For the python
side, any defs will appear in the ns variable. To enable python
imports into the global available modules, add a PyImports section to
the metadata that will add listed modules as imports.

Do some refactoring of the python side to shrink it below the 4K
ZipFile limit. This includes switching to 2 space indent (similar to
JS side), using dictionary comprehension in one place, using ternary
expressions in a couple of places to avoid multiline statements, etc.

Also, disable prints to stdout unless VERBOSE environment variable is
set. This squelches the logging of the event and new fragment for
every macro call when running locally.